refactor: rename capymoa.pi to capymoa.interval, fix import consistency
- Rename capymoa.pi -> capymoa.interval (avoids ambiguity with the
  math constant) and update tests, docs, and the notebook accordingly.
- Restore the explanatory comment on the core import in
  capymoa/__init__.py that got dropped when it changed from `stream`
  to `core`.
- Use the public `capymoa.core` import instead of reaching into the
  private `capymoa.core._instance` submodule in stream/_stream.py,
  stream/_csv_stream.py, stream/torch.py,
  stream/preprocessing/{pipeline,transformer}.py, and
  drift/detectors/abcd.py, for consistency with the rest of the
  codebase. Verified this doesn't reintroduce the circular import the
  private path was dodging, since capymoa/core/__init__.py binds
  Instance et al. before importing the submodules that pull
  capymoa.stream back in.
