+++ src/capymoa/interval/__init__.py
