+++ src/capymoa/interval/_adaptive_prediction_interval.py
