+++ src/capymoa/interval/_mean_and_standard_deviation_estimation.py
