+++ tests/test_interval.py
@@ -4,7 +4,7 @@ from capymoa.evaluation import (
 )
 from capymoa.datasets import Fried
 from capymoa.base import PredictionIntervalLearner
-from capymoa.pi import (
+from capymoa.interval import (
     MVE,
     AdaPI,
 )
